requirements.txt
pyspark==3.5.0
python-dotenv==1.0.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pymongo==4.6.1
pandas==2.1.1
numpy==1.26.0
cryptography==41.0.5
urllib3==2.0.7
dash==2.13.0
dash-bootstrap-components==1.5.0
plotly==5.17.0
flask==2.3.3